requirements.txt
Flask>=1.0.0
Flask-Cors==3.0.3
Flask-Limiter==1.0.1
Flask-Migrate==2.1.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
psycopg2==2.8.4
python-dotenv==0.10.3
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.32.14
urllib3>=1.24.2
Werkzeug>=0.15.3